Split refactor into new file, add local package structure
